+++ src/components/Transacrions/TransactionList/TransactionsItem.tsx
@@ -9,25 +9,27 @@ import {Transaction} from '../../../types';
 import {selectCategory} from '../../../store/categorySlice';
 import dayjs from 'dayjs';
 
-interface Props{
+interface Props {
   transaction: Transaction;
 }
 
 const TransactionsItem: React.FC<Props> = ({transaction}) => {
   const dispatch = useAppDispatch();
   const deleting = useAppSelector(selectDeletingTransaction);
-
   const categories = useAppSelector(selectCategory);
-  const category = categories.find(category=>category.id === transaction.category);
+  const category = categories.find(category => category.id === transaction.category);
 
   const removeTransaction = async (id: string) => {
-    try {
-      await dispatch(deleteTransaction(id));
-      toast.success('Transaction deleted');
-    } catch (error) {
-      toast.error('Could not delete transaction!');
-    } finally {
-      await dispatch(fetchTransactions());
+    const answer = confirm('Are you sure you want to delete this transaction?');
+    if (answer) {
+      try {
+        await dispatch(deleteTransaction(id));
+        toast.success('Transaction deleted');
+      } catch (error) {
+        toast.error('Could not delete transaction!');
+      } finally {
+        await dispatch(fetchTransactions());
+      }
     }
   };
   const showEditModal = (id: string) => {
@@ -39,7 +41,14 @@ const TransactionsItem: React.FC<Props> = ({transaction}) => {
         <Col className="pe-0">
           <Card.Body>
             <Card.Text>
-              <span>{dayjs(transaction.createdAt).format('DD.MM.YYYY HH:mm:ss')}</span>
+              <span>{dayjs(transaction.createdAt).format('DD.MM.YYYY HH:mm')}</span>
+            </Card.Text>
+          </Card.Body>
+        </Col>
+        <Col>
+          <Card.Body>
+            <Card.Text className="text-capitalize">
+              {category.title}
             </Card.Text>
           </Card.Body>
         </Col>
@@ -49,7 +58,7 @@ const TransactionsItem: React.FC<Props> = ({transaction}) => {
             "text-success text-capitalize"
             :
             "text-danger text-capitalize"}>
-            {category.type==='income' ?'+': '-'}{transaction.amount} KGS
+            {category.type === 'income' ? '+' : '-'}{transaction.amount} KGS
           </Card.Text>
         </Col>
         <Col xs={3} className="text-end">
@@ -57,6 +66,7 @@ const TransactionsItem: React.FC<Props> = ({transaction}) => {
                   className="mx-1"
                   onClick={() => showEditModal(transaction.id)}
           >edit</Button>
+
           <SpinnerBtn className="mx-1"
                       isSending={deleting === transaction.id}
                       variant="danger"
